// src/redux/employee.types.ts
export interface Employee {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

export interface EmployeeState {
  employees: Employee[];
  selectedEmployee: Employee | null;
  isEditing: boolean;
  isLoading: boolean;
  error: string | null;
}